test: add test suite targeting the C library directly
Spiel's API and language-bindings are feature-tested with Python,
but getting line coverage and performing sanitizer runs under these
conditions is difficult (or impossible).

Add an additional test suite with GLib tests, which can be run under a
sanitizer and produce better reports. The current tests will be skipped
if introspection it built, so the CI can perform sanitizer runs.